(refactor) relocate challenge API controller - move code to controller.js
